Fix onboarding event bugs
The onboarding event was trying to provide a recommendation for weather.com, when the weather.com add-on had been removed from the set of recommendations.

Additionally, the onboarding event now uses the waitForWindow method for showing the panel at the correct time, rather than trying to manually force open a panel.

Closes #46