Only call Rec.destroy if Rec != null
Previously Rec.destroy was being called on every call to cleanup function. The cleanup function is used in the uninstall process. When rec.destroy was called when Rec == null (set to null on first pass of cleanup function), the cleanup function would fail and the addon would not uninstall itself. The survey would then be displayed on the next startup, the cleanup function would fail again. There is now a check to only call Rec.destroy if it is not null (where if rec == null, rec.destroy has already been called)

Closes #77