Add stylesheet for button, remove mouseover / mouseout
The styles for the enhance button were moved into a css file that is applied using the sdk style sheet service. This means that the button styles no longer need to be applied with js, and in particular, :hover can be used, so that the mouseover / mouseout listeners can be removed from the js.