Resolve onboarding bug where wikipedia is opened on every startup
I'm using simple prefs to note that the onboarding event (where wikipedia is opened) has occurred. Previously the pref was set after the page had fully loaded, meaning that if users closed the wikipedia tab early (on a slow internet connection), the onboarding event would not be marked as complete. Therefore I've moved the pref set to as soon as the onboard method is called, eliminating this race condition.

I replicated the issue by simulating a very slow internet connection on my computer and testing the addon. This PR fixes the issue in my testing.